Moved testing class into server general testing project and removed redundant dependencies
