+++ server/authentication/build.gradle
@@ -3,13 +3,8 @@ apply plugin: 'application'
 mainClassName = 'AuthenticationServiceMain'
 
 dependencies {
-    implementation project(':core')
-
     implementation 'io.vertx:vertx-jdbc-client-scala_2.12:3.5.2'
-
-    implementation 'io.vertx:vertx-web-scala_2.12:3.5.2'
     
     testImplementation project(':server:testing')
-    testImplementation 'io.vertx:vertx-unit:3.5.2'
     testImplementation group: 'org.hsqldb', name: 'hsqldb', version: '2.0.0'
 }
