+++ server/authentication/src/test/scala/it/cwmp/authentication/AuthenticationServiceTest.scala
@@ -1,6 +1,6 @@
 package it.cwmp.authentication
 
-import it.cwmp.utils.VerticleTesting
+import it.cwmp.testing.VerticleTesting
 import org.scalatest.Matchers
 
 import scala.concurrent.Promise
